Enable collations in the XPath static context
Allow setting collations in the static context an XPath Compiler uses to
populate the static context of compiled XPaths.

This takes the approach of using a new StaticContext object that we use
to set the static context on the underlying Saxon XPathCompiler. This
separation allows us to treat both the context and the XPathCompiler as
immutable, which simplifies a bunch of thread-safety issues with
XPathCompilers (which is, they're safe if they're unmodified...)