Update collations code
Update collations code in the XPath stuff to reflect my improved
understanding of how it works in Saxon and the removal of methods for
declaring custom collations on XPath, XSLT, and XQuery compilers, which
were deprecated back in 9.5 or 9.6 but not documented.

Now custom collations (which are needed much more rarely than I had
though) should be declared on the `Saxon::Processor` itself, and are
shared by all resources that use the Processor.

This also allows us to move to Saxon 9.9 without test failures...