Add support for passing in XSLT parameters
Support the full range of static, global, initial (non-tunneling), and
initial (tunneling).

As part of this, change the name of `XSLT::StaticContext` to
`EvaluationContext` to reflect that it's used for much more than the
static compile-time stuff.

Refactor QName to pull in the ruby-string-to-qname resolution code that
was in the XPath `StaticContext`, since this is really where it belongs.

There's some tire-kicking to be done to figure out whether this API
works the way I want, so this is definitely transitional.