Add support for more XDM types
Add handling for Names, NCNames, Booleans, and QNames

Booleans require more explicit separation of inferred-item-type style
creation and explicit lexical string-style creation because of the whole
`'false'` is truthy (because all strings are truthy) issue.

QNames are a whole other problem, because they're namespace-sensitive
and therefore have no meaningful lexical string representation outside
an XML doc - the prefix being the one part of a QName you don't actually
need, and the URI, which you do, being the one part not present in a
straight lexical string representation, as used within an XML doc...

QNames get some special case handling, which we'll need to revisit when
we get to xs:NOTATION, the other namespace-sensitive simple XSD type.