Add Saxon.XML() convenience function
Add `XML()` convenience function to match `saxon-xslt` and Nokogiri.

Allow explicit setting of character encoding on inputs for situations
where the input encoding is not specified in a way that the parser will
detect, or is specified wrongly and must be overridden.

Big rewrite of README.

Improve API documentation slightly. (Document a few undocumented modules
and methods.)