Add SequenceType class
This wraps the `s9api.SequenceType` class. The main place this concept
is used currently is in XPath variable declaration, where the underlying
Java API uses the ItemType and OccurrenceIndicator directly without
using the SequenceType (it is used in XSLT/XPath extension function
definition which I haven't gotten to yet).

Next step is to use our wrapper class to neaten up the variable
declaration DSL.