Add Feature flags code to cope with other Saxons
Some things work on Saxon 9.9 but not 9.8, or on 10.0 but not 9.x.

In order to support that we need a way to report the version of the
loaded underlying Saxon, and a way to (for now, at least) switch a
method off for certain Saxon versions.

Move the Saxon::VERSION constant to Saxon::Version::WRAPPER because we
now have parallel versions to consider - the Ruby wrapper and the Java
library.