Make `require 'saxon-rb'` work again
requiring 'saxon-rb' was a no-op for ages, then I made it work, and then
took it away while hunting the real source of the
can't-use-own-saxon-jar problems (it was not the root cause). Since some
people are using `require 'saxon-rb'` and it doesn't cause problems any
more, let's make it work again.